+++ examples/local_multiplayer.rs
@@ -6,11 +6,12 @@ use std::f32::consts::FRAC_PI_4;
 
 use bevy::{
     color::palettes::tailwind::{BLUE_600, RED_600},
+    input::gamepad::{GamepadConnection, GamepadConnectionEvent},
     prelude::*,
 };
 use bevy_enhanced_input::prelude::*;
 
-use player_box::{PlayerBox, PlayerBoxBundle, PlayerBoxPlugin, PlayerColor, DEFAULT_SPEED};
+use player_box::{PlayerBox, PlayerBoxPlugin, PlayerColor, DEFAULT_SPEED};
 
 fn main() {
     App::new()
@@ -28,32 +29,29 @@ struct GamePlugin;
 impl Plugin for GamePlugin {
     fn build(&self, app: &mut App) {
         app.add_input_context::<PlayerBox>()
+            .add_observer(Self::apply_movement)
+            .add_observer(Self::rotate)
             .add_systems(Startup, Self::spawn)
-            .observe(Self::apply_movement)
-            .observe(Self::rotate);
+            .add_systems(Update, Self::update_gamepads);
     }
 }
 
 impl GamePlugin {
     fn spawn(mut commands: Commands) {
-        commands.spawn(Camera2dBundle::default());
+        commands.spawn(Camera2d);
 
         // Spawn two players with different assigned indices.
         commands.spawn((
-            PlayerBoxBundle {
-                transform: Transform::from_translation(Vec3::X * 50.0),
-                color: PlayerColor(RED_600.into()),
-                ..Default::default()
-            },
-            PlayerIndex(0),
+            PlayerBox,
+            Transform::from_translation(Vec3::X * 50.0),
+            PlayerColor(RED_600.into()),
+            Player::First,
         ));
         commands.spawn((
-            PlayerBoxBundle {
-                transform: Transform::from_translation(-Vec3::X * 50.0),
-                color: PlayerColor(BLUE_600.into()),
-                ..Default::default()
-            },
-            PlayerIndex(1),
+            PlayerBox,
+            Transform::from_translation(-Vec3::X * 50.0),
+            PlayerColor(BLUE_600.into()),
+            Player::Second,
         ));
     }
 
@@ -67,43 +65,73 @@ impl GamePlugin {
         let mut transform = players.get_mut(trigger.entity()).unwrap();
         transform.rotate_z(FRAC_PI_4);
     }
+
+    fn update_gamepads(
+        mut commands: Commands,
+        mut connect_events: EventReader<GamepadConnectionEvent>,
+        mut gamepads: ResMut<Gamepads>,
+    ) {
+        for event in connect_events.read() {
+            match event.connection {
+                GamepadConnection::Connected(_) => gamepads.push(event.gamepad),
+                GamepadConnection::Disconnected => {
+                    if let Some(index) = gamepads.iter().position(|&entity| entity == event.gamepad)
+                    {
+                        gamepads.swap_remove(index);
+                    }
+                }
+            }
+        }
+
+        // Trigger contexts rebuild to update associated gamepads.
+        commands.trigger(RebuildInputContexts);
+    }
 }
 
-#[derive(Component, Deref)]
-struct PlayerIndex(usize);
+#[derive(Component, Clone, Copy, PartialEq, Eq, Hash)]
+enum Player {
+    First,
+    Second,
+}
+
+/// A resource that tracks all connected gamepads to pick them by index.
+#[derive(Resource, Default, Deref, DerefMut)]
+struct Gamepads(Vec<Entity>);
 
 impl InputContext for PlayerBox {
     fn context_instance(world: &World, entity: Entity) -> ContextInstance {
+        let mut ctx = ContextInstance::default();
+
         // Could be stored in the context itself, but it's usually
         // better to have a separate component that is shared
         // across all contexts.
-        let index = **world.get::<PlayerIndex>(entity).unwrap();
+        let player = *world.get::<Player>(entity).unwrap();
 
         // By default context read inputs from all gamepads,
         // but for local multiplayer we need assign specific
         // gamepad index.
-        let mut ctx = ContextInstance::with_gamepad(index);
+        let gamepads = world.resource::<Gamepads>();
+        if let Some(&entity) = gamepads.get(player as usize) {
+            ctx.set_gamepad(entity);
+        }
 
         // Assign different mappings based player index.
-        match index {
-            0 => {
+        match player {
+            Player::First => {
                 ctx.bind::<Move>()
                     .with_wasd()
                     .with_stick(GamepadStick::Left);
                 ctx.bind::<Rotate>()
                     .with(KeyCode::Space)
-                    .with(GamepadButtonType::South);
+                    .with(GamepadButton::South);
             }
-            1 => {
+            Player::Second => {
                 ctx.bind::<Move>()
                     .with_arrows()
                     .with_stick(GamepadStick::Left);
                 ctx.bind::<Rotate>()
                     .with(KeyCode::Numpad0)
-                    .with(GamepadButtonType::South);
-            }
-            _ => {
-                panic!("game expects only 2 players");
+                    .with(GamepadButton::South);
             }
         }