+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Rename `Press` into `JustPress`.
 - Rename `Down` into `Press` to avoid collision with `Down` from `bevy_picking`.
+- Replace `ContextInstance::with_gamepad` builder with just `ContextInstance::set_gamepad` setter.
 
 ## [0.2.0] - 2024-11-03